refactor(cardInformation): 상수 파일명 수정
