+++ hooks/src/lib/constants/cardInformation.ts
